requirements.txt
keyboard
pynput
pydivert
vgamepad
pillow
customtkinter